Add some tests
This commits adds a test to check if the fuzzer can
find a crash, and to make sure that it doesn't find
anything when the target isn't crashing.

We should probably refactor a bit the fuzzer.py file
to make it expose a better API, instead of relying on `logging`.